% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseB.intern.R
\name{chooseB.intern}
\alias{chooseB.intern}
\title{Tune the number of iterations for variable selection using varselbest}
\usage{
chooseB.intern(
  res.varselbest,
  gridB = NULL,
  xlim = NULL,
  plotvar = NULL,
  cex = 0.2,
  type = "b",
  xlab = "B",
  ylab = "proportion",
  graph = TRUE,
  pch = 16
)
}
\arguments{
\item{res.varselbest}{an output from the varselbest function}

\item{gridB}{a grid for the number of iterations. By default, the grid is tuned to 1:B where B is the argument used in varselbest.}

\item{xlim}{the x limits (x1, x2) of the plot}

\item{plotvar}{index of variables for which a curve is ploted}

\item{cex}{a numerical vector giving the amount by which plotting characters and symbols should be scaled relative to the default}

\item{type}{what type of plot should be drawn}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{graph}{a boolean. If FALSE, no graphics are ploted. Default value is TRUE}

\item{pch}{an integer}
}
\description{
Tune the number of iterations for variable selection using varselbest
}
\references{
Bar-Hen, A. and Audigier, V. An ensemble learning method for variable selection: application to high dimensional data and missing values. ArXiv e-prints <arXiv:1808.06952>
}
\keyword{internal}
