% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initfastnmf.intern.R
\name{initfastnmf}
\alias{initfastnmf}
\title{initialize fastnmf}
\usage{
initfastnmf(
  method,
  Mtilde,
  listpart = NULL,
  nb.clust = NULL,
  parameter.kmeans = list(nstart = 100, iter.max = 50, algorithm = c("Hartigan-Wong",
    "Lloyd", "Forgy", "MacQueen"), trace = FALSE),
  parameter.minibatchkmeans = list(batch_size = 10, num_init = 1, max_iters = 50,
    init_fraction = 1, initializer = "kmeans++", early_stop_iter = 10, verbose = FALSE,
    CENTROIDS = NULL, tol = 1e-04, tol_optimal_init = 0.3, seed = 1)
)
}
\arguments{
\item{method}{a vector giving initialisation methods among "BOK", "kmeans", "minibatchkmeans" "sample"}

\item{Mtilde}{a fuzzy connectivity matrix}

\item{listpart}{a list of partitions (required for kmeans and minibatchkmeans method)}

\item{parameter.kmeans}{list of arguments for kmeans function}

\item{parameter.minibatchkmeans}{list of arguments for MiniBatchKmeans function}
}
\description{
initialize fastnmf
}
\keyword{internal}
