% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclustboot.intern.R
\name{mclustboot.intern}
\alias{mclustboot.intern}
\title{MclustBootstrap with nboot = 1 and the same output as Mclust}
\usage{
mclustboot.intern(don, G = NULL, modelNames = NULL, verbose = FALSE)
}
\arguments{
\item{don}{matrix, or data frame}

\item{G}{An integer vector specifying the numbers of mixture components}

\item{modelNames}{A vector of character strings indicating the models to be fitted}

\item{verbose}{A logical controlling if a text progress bar is displayed}
}
\description{
MclustBootstrap with nboot = 1 and the same output as Mclust
}
\keyword{internal}
