% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myem.mix.R
\name{myem.mix}
\alias{myem.mix}
\title{internal function}
\usage{
myem.mix(dataset, K, boot = FALSE, silent = TRUE, nbquali = 0, eps = 1e-04)
}
\description{
prend en entrée un jeu de données, ajoute une variable classe, et renvoie les parametre du general location model. Une des particularité est l'initialisation du EM par un modèle de mélange. En effet em.mix a sa propre façon d'initialiser qui parfois renvoie des paramètres identiques pour plusieurs groupes
}
\keyword{internal}
