% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefold.chooser.R
\name{onefold.chooser}
\alias{onefold.chooser}
\title{one fold cross-validation for specifying threshold r}
\usage{
onefold.chooser(
  data.train.k,
  data.test.k,
  jj,
  grid.r,
  nb.clust,
  nnodes,
  sizeblock,
  method.select,
  B,
  modelNames,
  K,
  path.outfile,
  nbvarused
)
}
\arguments{
\item{data.train.k}{a train set}

\item{data.test.k}{a test set}

\item{jj}{name of the outcome used for variable selection}

\item{grid.r}{a grid for the tuning parameter r}

\item{nb.clust}{number of clusters}

\item{sizeblock}{number of sampled variables at each iteration}

\item{method.select}{variable selection method}

\item{B}{number of iterations}

\item{modelNames}{mixture model specification for imputation of subsets}

\item{K}{number of fold}

\item{path.outfile}{a path for message redirection}

\item{nbvarused}{a maximal number of selected variables (can be required with a large number of variables)}
}
\description{
one fold cross-validation for specifying threshold r
}
\keyword{internal}
