% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silhouette.intern.R
\name{Silhouette.intern}
\alias{Silhouette.intern}
\title{Compute Silhouette index}
\usage{
Silhouette.intern(d, cl)
}
\arguments{
\item{d}{a distance matrix}

\item{cl}{a vector of integer defining the partition}
}
\value{
a real
}
\description{
Compute Silhouette index
}
\keyword{internal}
